requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
python-dotenv
google-genai